+++ memi/memi-web/src/main/java/uk/ac/ebi/interpro/metagenomics/memi/springmvc/modelbuilder/HomePageViewModelBuilder.java
@@ -77,9 +77,9 @@ public class HomePageViewModelBuilder extends AbstractBiomeViewModelBuilder<Home
         final int publicRunCount = runDAO.countAllPublic();
         final int privateRunCount = runDAO.countAllPrivate();
 
-        final Map<String, Integer> experimentCountMap = runDAO.retrieveRunCountsGroupedByExperimentType(3);
-        final Map<String, Integer> transformedExperimentCountMap = transformMap(experimentCountMap);
-        final Integer numOfDataSets = getNumOfDataSets(experimentCountMap);
+        final Map<String, Long> experimentCountMap = runDAO.retrieveRunCountsGroupedByExperimentType(3);
+        final Map<String, Long> transformedExperimentCountMap = transformMap(experimentCountMap);
+        final Long numOfDataSets = getNumOfDataSets(experimentCountMap);
 
         List<Study> studies = null;
         // If case: if nobody is logged in
@@ -134,8 +134,8 @@ public class HomePageViewModelBuilder extends AbstractBiomeViewModelBuilder<Home
         }
     }
 
-    private Map<String, Integer> transformMap(Map<String, Integer> experimentCountMap) {
-        Map<String, Integer> result = new TreeMap<String, Integer>(
+    private Map<String, Long> transformMap(Map<String, Long> experimentCountMap) {
+        Map<String, Long> result = new TreeMap<String, Long>(
                 //Comparator is tested in HomePageSamplesComparatorTest
                 new Comparator<String>() {
                     @Override
@@ -155,7 +155,7 @@ public class HomePageViewModelBuilder extends AbstractBiomeViewModelBuilder<Home
 
                 });
         for (String key : experimentCountMap.keySet()) {
-            Integer value = experimentCountMap.get(key);
+            Long value = experimentCountMap.get(key);
             if (key.equalsIgnoreCase("assembly")) {
                 result.put("assemblies", value);
             } else {
@@ -165,8 +165,8 @@ public class HomePageViewModelBuilder extends AbstractBiomeViewModelBuilder<Home
         return result;
     }
 
-    private Integer getNumOfDataSets(Map<String, Integer> experimentCountMap) {
-        Integer result = 0;
+    private Long getNumOfDataSets(Map<String, Long> experimentCountMap) {
+        Long result = 0L;
         for (String key : experimentCountMap.keySet()) {
             result += experimentCountMap.get(key);
         }
